Re-structure project commons into a separate static library Demo program source file is now identified as such, in preparation for multiple programs for different OS'es (and cross-platform demo mode)
