+++ demo.cpp
@@ -5,22 +5,10 @@
 
 #include <SFML/Window.hpp>
 
-#include "main.hpp"
-#include "HexagoScreenSaver.hpp"
-#include "argument_parser.hpp"
+#include "hexago/hexago.hpp"
+#include "hexago/HexagoScreenSaver.hpp"
+#include "hexago/argument_parser.hpp"
 
-
-// Version numbers are passed as preprocessor definitions by CMake
-const hexago::version_t hexago::VERSION = {
-    HEXAGO_VERSION_MAJOR,
-    HEXAGO_VERSION_MINOR,
-    HEXAGO_VERSION_PATCH,
-    HEXAGO_VERSION_BUILD,
-    HEXAGO_VERSION_STRING,
-};
-
-// program copyright notice
-const char* copyright = "Copyright (c) 2017 Joshua Saxby. All rights reserved.";
 
 int main(int argc, char* argv[]) {
     // build window title / console 'greet' message
@@ -29,7 +17,7 @@ int main(int argc, char* argv[]) {
     window_title_stream << " (build " << hexago::VERSION.build << ")";
     std::string window_title = window_title_stream.str();
     // print Hexago version and copyright notice, on separate lines
-    printf("%s\n%s\n", window_title.c_str(), copyright);
+    printf("%s\n%s\n", window_title.c_str(), hexago::copyright);
     // get supported fullscreen video modes
     std::vector<sf::VideoMode> video_modes = sf::VideoMode::getFullscreenModes();
     // window settings
