+++ hexago/Hexagon.cpp
