+++ hexago/HexagonFactory.cpp
