+++ hexago/HexagonFactory.hpp
