+++ hexago/HexagonFactoryConfig.cpp
