+++ hexago/HexagonFactoryConfig.hpp
